Change mocha to tape.
